+++ plone/app/widgets/tests/test_dx.py
@@ -1132,6 +1132,21 @@ class RichTextWidgetTests(unittest.TestCase):
             base_args['pattern_options']['upload']['relativePath'],
             '@@fileUpload')
 
+        self.assertEqual(
+            base_args['pattern_options']['relatedItems']['mode'],
+            'browse'
+        )
+        self.assertEqual(
+            base_args['pattern_options']['relatedItems']['basePath'],
+            '/plone'
+        )
+        # Next, we're only testing for the existance of folderTypes parameter.
+        # Values can either be set via GenericSetup (tinymce.xml) or
+        # plone.app.registry.
+        self.assertTrue(
+            'folderTypes' in base_args['pattern_options']['relatedItems']
+        )
+
         if not PLONE50:
             self.assertEqual(base_args['pattern_options']['anchorSelector'],
                              self.portal.portal_tinymce.anchor_selector)